PCX-785: Padding and logo updates
