--- payment/src/main/res/layout/list_item_logo.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<ImageView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/image_logo"
-    android:layout_width="@dimen/pmlogo_width"
-    android:layout_height="@dimen/pmlogo_height"
-    android:adjustViewBounds="true"
-    android:contentDescription="@string/pmpage_contentdescription" />
