+++ payment/src/main/res/layout/list_item_networkcard.xml
@@ -23,35 +23,25 @@
                 android:layout_alignParentTop="true"
                 android:clickable="true"
                 android:focusable="true"
-                android:paddingStart="@dimen/pmborder_xsmall"
-                android:paddingEnd="@dimen/pmborder_xsmall"
+                android:paddingTop="@dimen/pmborder_xsmall"
+                android:paddingBottom="@dimen/pmborder_xsmall"
                 android:gravity="center_vertical"
                 android:orientation="horizontal">
 
-                <include layout="@layout/view_logos" />
-
-                <TextSwitcher
-                    android:id="@+id/textswitcher_title"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="@dimen/pmborder_small"
-                    android:inAnimation="@anim/fade_in"
-                    android:outAnimation="@anim/fade_out">
+                <ImageView
+                    android:id="@+id/image_logo"
+                    android:layout_width="@dimen/pmlogo_width"
+                    android:layout_height="@dimen/pmlogo_height"
+                    android:adjustViewBounds="true"
+                    android:layout_marginEnd="@dimen/pmborder_medium"
+                    android:contentDescription="@string/pmpage_contentdescription" />    
 
                     <TextView
-                        android:id="@+id/title0"
+                        android:id="@+id/text_title"
                         android:textAppearance="?attr/textAppearanceSubtitle1"
                         android:textColor="?attr/colorOnSurface"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content" />
-
-                    <TextView
-                        android:id="@+id/title1"
-                        android:textAppearance="?attr/textAppearanceSubtitle1"
-                        android:textColor="?attr/colorOnSurface"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content" />
-                </TextSwitcher>
             </LinearLayout>
 
             <LinearLayout
@@ -59,9 +49,8 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_below="@id/layout_header"
-                android:orientation="vertical"
                 android:paddingTop="@dimen/pmborder_medium"
-                android:paddingEnd="@dimen/pmborder_xsmall"
+                android:orientation="vertical"
                 android:visibility="gone" />
         </RelativeLayout>
