--- payment/src/main/res/layout/view_logos.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<TableLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/tablelayout_logo"
-    android:layout_width="@dimen/pmcard_logos_portrait"
-    android:layout_height="wrap_content" />
