+++ payment/src/main/res/layout/widget_checkbox.xml
@@ -5,9 +5,7 @@
     android:layout_height="wrap_content"
     android:gravity="center_vertical"
     android:orientation="horizontal"
-    android:paddingStart="@dimen/pmborder_xsmall"
-    android:paddingTop="@dimen/pmborder_medium"
-    android:paddingBottom="@dimen/pmwidget_padding">
+    android:layout_marginBottom="@dimen/pmborder_xlarge">
 
     <LinearLayout
         android:layout_width="0dp"
