+++ payment/src/main/res/values/dimens.xml
@@ -3,10 +3,10 @@
     <dimen name="pmborder_xsmall">8dp</dimen>
     <dimen name="pmborder_small">12dp</dimen>
     <dimen name="pmborder_medium">16dp</dimen>
+    <dimen name="pmborder_xmedium">20dp</dimen>
     <dimen name="pmborder_large">24dp</dimen>
-    <dimen name="pmborder_xlarge">30dp</dimen>
+    <dimen name="pmborder_xlarge">32dp</dimen>
 
-    <dimen name="pmwidget_padding">4dp</dimen>
     <dimen name="pmcard_elevation">0dp</dimen>
     <dimen name="pmcard_stroke">1dp</dimen>
     <dimen name="pmcard_logos_portrait">108dp</dimen>
